requirements.txt
git+https://github.com/Rapptz/discord.py
PyNaCl
dnspython
async-timeout
typing
discord-pretty-help
praw